requirements/requirements.dev.txt
Django==2.0.13
djangorestframework==3.9.1
django-cors-headers==2.4.0
django-cors-middleware==1.3.1
django-filter==2.0.0
psycopg2-binary==2.7.4
django-jet==1.0.7
requests==2.20.0
oic==0.14.0
Pillow==5.3.0
django-ckeditor==5.7.1
uwsgi==2.0.18
whitenoise==4.0
ipython==7.7.0
django-mptt==0.9.1
